requirements.txt
typing-extensions
langchain-openai
langchain-experimental
pandas
yfinance
praw
feedparser
stockstats
eodhd
langgraph
chromadb
setuptools
backtrader
akshare
tushare
finnhub-python
parsel
requests
tqdm
pytz
pyyaml>=6.0
kerykeion>=4.14.0
hyperliquid-python-sdk
eth-account>=0.10.0
redis
chainlit
rich
questionary
langchain_anthropic
langchain-google-genai
